homepage basics
created basic homepage, noteable things missing so far with these two
pages: currently all link and nav features are placeholders. Also the
calendar widget i found in 2 seconds and is pretty garbage so we will
have to look for a better solution for the final implementation of this.
Again css is very basic no color scheme yet - Colin