Dynamic generation issue
This page is going to be an issue currently because it will need to
generate based on information in the database, in other words how ever
many classes are missing that many rows need to be created in the html.
Going to require PHP code or something like that. Also, the
javascript/jquery will be an issue because of this dynamic generation.